created function do display experience and implement in detail
Removed html from detail.php relating to displaying work experience and moved it to the function displayExperience. Instead of using $_get every time for information in the array, one call for the entire array is saved as a variable and passed as a single parameter to the function and returns every work experience item in html.